addition of 'patch' route, checks for the existence of data in the body and in the database and processing the date using the 'date-fns' library

+++ 01-desafio/routes.js
@@ -1,19 +1,19 @@
 import { randomUUID } from "crypto";
 import { Database } from "./database.js";
 import { buildRoutePath } from "./utils/build-route-path.js";
+import { format } from "date-fns";
+import { ptBR } from "date-fns/locale";
 
 const db = new Database();
 
-function formatDate(date = new Date(), format = "dd/mm/aaaa") {
-  const map = {
-    mm: date.getMonth() + 1,
-    dd: date.getDate(),
-    aa: date.getFullYear().toString().slice(-2),
-    aaaa: date.getFullYear(),
-  };
-
-  return format.replace(/mm|dd|aa|aaaa/gi, (matched) => map[matched]);
+function taskDateFormatted() {
+  let date = new Date();
+  let taskDateTodayFormatted = format(date, "dd 'de' LLLL 'de' y 'às' HH:mm'h'", {
+    locale: ptBR,
+  });
+  return taskDateTodayFormatted;
 }
+
 
 export const routes = [
   {
@@ -36,8 +36,8 @@ export const routes = [
     handler: (request, response) => {
       const { title, description } = request.body;
       const completed_at = null;
-      const updated_at = null;
-      const created_at = formatDate();
+      const updated_at = taskDateFormatted();
+      const created_at = taskDateFormatted();
       const task = {
         id: randomUUID(),
         title,
@@ -49,7 +49,9 @@ export const routes = [
 
       db.insert("tasks", task);
 
-      return response.writeHead(201).end(JSON.stringify({ msg: "Tarefa inserida com sucesso" }));
+      return response
+        .writeHead(201)
+        .end(JSON.stringify({ msg: "Tarefa inserida com sucesso" }));
     },
   },
   {
@@ -58,9 +60,19 @@ export const routes = [
     handler: (request, response) => {
       const { id } = request.params;
 
+      //validação se o id pertence a uma task salva no banco de dados.
+      const [task] = db.select("tasks", { id });
+      if (!task) {
+        return response
+          .writeHead(404)
+          .end(JSON.stringify({ msg: "id não encontrado" }));
+      }
+
       db.delete("tasks", id);
 
-      return response.writeHead(204).end();
+      return response
+        .writeHead(200)
+        .end(JSON.stringify({ msg: "task deletada com sucesso" }));
     },
   },
   {
@@ -74,30 +86,55 @@ export const routes = [
       const { title, description } = request.body;
 
       // verificar se title e description foram retornados
-      if (!title || !description) {
-        return response.writeHead(400).end(JSON.stringify({ msg: "falta parâmetro, título ou descrição" }));
+      if (!title && !description) {
+        return response
+          .writeHead(400)
+          .end(JSON.stringify({ msg: "falta parâmetro, título ou descrição" }));
       }
 
-      // verificar se os dados já existem
-      const [task] = db.select('tasks', { id })
-
+      //validação se o id pertence a uma task salva no banco de dados.
+      const [task] = db.select("tasks", { id });
       // responder case task não seja encontrada
       if (!task) {
-        return response.writeHead(404).end(JSON.stringify({ msg: "task não encontrada" })) 
+        return response
+          .writeHead(404)
+          .end(JSON.stringify({ msg: "task não encontrada" }));
       }
-
 
       // chave: valor (body) e valor (select)
       // operador de coalescência
-      db.update("tasks", id, {
-        title: title ?? task.title, 
+      db.update("tasks", id, { ...task,
+        title: title ?? task.title,
         description: description ?? task.description,
-        completed_at: task.completed_at,
-        created_at: task.created_at,
-        updated_at: new Date(),
-      })
+        updated_at: taskDateFormatted(),
+      });
 
-      return response.writeHead(200).end(JSON.stringify({ msg: "atualizado com sucesso" }));
+      return response
+        .writeHead(200)
+        .end(JSON.stringify({ msg: "atualizado com sucesso" }));
+    },
+  },
+  {
+    method: "PATCH",
+    path: buildRoutePath("/tasks/:id/complete"),
+    handler: (request, response) => {
+      const { id } = request.params;
+
+      //validação se o id pertence a uma task salva no banco de dados.
+      const [task] = db.select("tasks", { id });
+      if (!task) {
+        return response
+          .writeHead(404)
+          .end(JSON.stringify({ msg: "id não encontrado" }));
+      }
+
+      // capturar valor booleano: "completed_at"
+      const isTaskCompleted = !!task.completed_at;
+      const completed_at = isTaskCompleted ? null : taskDateFormatted();
+
+      db.update("tasks", id, { ...task, completed_at });
+
+      return response.writeHead(204).end();
     },
   },
 ];
